AVRO-2842: Add PSR4 and code formatting.
