+++ lang/php/lib/DataFile/AvroDataIOException.php
@@ -17,15 +17,14 @@
  * limitations under the License.
  */
 
-include __DIR__ . '/../vendor/autoload.php';
+namespace Apache\Avro\DataFile;
 
-define('AVRO_TEST_HELPER_DIR', __DIR__);
+use Apache\Avro\AvroException;
 
-define('TEST_TEMP_DIR', implode(DIRECTORY_SEPARATOR, [AVRO_TEST_HELPER_DIR, 'tmp']));
-
-define('AVRO_BASE_DIR', dirname(AVRO_TEST_HELPER_DIR, 3));
-define('AVRO_SHARE_DIR', implode(DIRECTORY_SEPARATOR, [AVRO_BASE_DIR, 'share']));
-define('AVRO_BUILD_DIR', implode(DIRECTORY_SEPARATOR, [AVRO_BASE_DIR, 'build']));
-define('AVRO_BUILD_DATA_DIR', implode(DIRECTORY_SEPARATOR, [AVRO_BUILD_DIR, 'interop', 'data']));
-define('AVRO_TEST_SCHEMAS_DIR', implode(DIRECTORY_SEPARATOR, [AVRO_SHARE_DIR, 'test', 'schemas']));
-define('AVRO_INTEROP_SCHEMA', implode(DIRECTORY_SEPARATOR, [AVRO_TEST_SCHEMAS_DIR, 'interop.avsc']));
+/**
+ * Raised when something unkind happens with respect to AvroDataIO.
+ * @package Avro
+ */
+class AvroDataIOException extends AvroException
+{
+}
