+++ lang/php/test/DataFileTest.php
@@ -17,184 +17,166 @@
  * limitations under the License.
  */
 
-require_once('test_helper.php');
+namespace Apache\Avro\Tests;
 
-class DataFileTest extends PHPUnit\Framework\TestCase
+use Apache\Avro\DataFile\AvroDataIO;
+use PHPUnit\Framework\TestCase;
+
+class DataFileTest extends TestCase
 {
-  private $data_files;
-  const REMOVE_DATA_FILES = true;
+    const REMOVE_DATA_FILES = true;
+    private $data_files;
 
-  static function current_timestamp() { return strftime("%Y%m%dT%H%M%S"); }
+    public function test_write_read_nothing_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-nothing-null-%s.avr', $codec));
+            $writers_schema = '"null"';
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            $dw->close();
 
-  protected function add_data_file($data_file)
-  {
-    if (is_null($this->data_files))
-      $this->data_files = array();
-    $data_file = "$data_file.".self::current_timestamp();
-    $full = join(DIRECTORY_SEPARATOR, array(TEST_TEMP_DIR, $data_file));
-    $this->data_files []= $full;
-    return $full;
-  }
-
-  protected static function remove_data_file($data_file)
-  {
-    if (file_exists($data_file))
-      unlink($data_file);
-  }
-
-  protected function remove_data_files()
-  {
-    if (self::REMOVE_DATA_FILES
-        && !empty($this->data_files))
-      foreach ($this->data_files as $data_file)
-        self::remove_data_file($data_file);
-  }
-
-  protected function setUp(): void
-  {
-    if (!file_exists(TEST_TEMP_DIR))
-      mkdir(TEST_TEMP_DIR);
-    $this->remove_data_files();
-  }
-
-  protected function tearDown(): void
-  {
-    $this->remove_data_files();
-  }
-
-  public function test_write_read_nothing_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-nothing-null-%s.avr', $codec));
-      $writers_schema = '"null"';
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      $dw->close();
-
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $datum = array_shift($read_data);
-      $dr->close();
-      $this->assertEquals(null, $datum);
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $datum = array_shift($read_data);
+            $dr->close();
+            $this->assertEquals(null, $datum);
+        }
     }
-  }
 
-  public function test_write_read_null_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-null-%s.avr', $codec));
-      $writers_schema = '"null"';
-      $data = null;
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      $dw->append($data);
-      $dw->close();
-
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $datum = array_shift($read_data);
-      $dr->close();
-      $this->assertEquals($data, $datum);
+    protected function add_data_file($data_file)
+    {
+        if (is_null($this->data_files)) {
+            $this->data_files = array();
+        }
+        $data_file = "$data_file." . self::current_timestamp();
+        $full = join(DIRECTORY_SEPARATOR, array(TEST_TEMP_DIR, $data_file));
+        $this->data_files [] = $full;
+        return $full;
     }
-  }
 
-  public function test_write_read_string_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-str-%s.avr', $codec));
-      $writers_schema = '"string"';
-      $data = 'foo';
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      $dw->append($data);
-      $dw->close();
-
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $datum = array_shift($read_data);
-      $dr->close();
-      $this->assertEquals($data, $datum);
+    public static function current_timestamp()
+    {
+        return strftime("%Y%m%dT%H%M%S");
     }
-  }
 
-  public function test_write_read_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-int-%s.avr', $codec));
-      $writers_schema = '"int"';
-      $data = 1;
+    public function test_write_read_null_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-null-%s.avr', $codec));
+            $writers_schema = '"null"';
+            $data = null;
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            $dw->append($data);
+            $dw->close();
 
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      $dw->append(1);
-      $dw->close();
-
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $datum = array_shift($read_data);
-      $dr->close();
-      $this->assertEquals($data, $datum);
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $datum = array_shift($read_data);
+            $dr->close();
+            $this->assertEquals($data, $datum);
+        }
     }
-  }
 
-  public function test_write_read_true_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-true-%s.avr', $codec));
-      $writers_schema = '"boolean"';
-      $datum = true;
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      $dw->append($datum);
-      $dw->close();
+    public function test_write_read_string_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-str-%s.avr', $codec));
+            $writers_schema = '"string"';
+            $data = 'foo';
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            $dw->append($data);
+            $dw->close();
 
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $read_datum = array_shift($read_data);
-      $dr->close();
-      $this->assertEquals($datum, $read_datum);
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $datum = array_shift($read_data);
+            $dr->close();
+            $this->assertEquals($data, $datum);
+        }
     }
-  }
 
-  public function test_write_read_false_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-false-%s.avr', $codec));
-      $writers_schema = '"boolean"';
-      $datum = false;
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      $dw->append($datum);
-      $dw->close();
+    public function test_write_read_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-int-%s.avr', $codec));
+            $writers_schema = '"int"';
+            $data = 1;
 
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $read_datum = array_shift($read_data);
-      $dr->close();
-      $this->assertEquals($datum, $read_datum);
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            $dw->append(1);
+            $dw->close();
+
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $datum = array_shift($read_data);
+            $dr->close();
+            $this->assertEquals($data, $datum);
+        }
     }
-  }
 
-  public function test_write_read_int_array_round_trip()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-wr-int-ary-%s.avr', $codec));
-      $writers_schema = '"int"';
-      $data = array(10, 20, 30, 40, 50, 60, 70);
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      foreach ($data as $datum)
-        $dw->append($datum);
-      $dw->close();
+    public function test_write_read_true_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-true-%s.avr', $codec));
+            $writers_schema = '"boolean"';
+            $datum = true;
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            $dw->append($datum);
+            $dw->close();
 
-      $dr = AvroDataIO::open_file($data_file);
-      $read_data = $dr->data();
-      $dr->close();
-      $this->assertEquals($data, $read_data,
-                          sprintf("in: %s\nout: %s",
-                                  json_encode($data), json_encode($read_data)));
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $read_datum = array_shift($read_data);
+            $dr->close();
+            $this->assertEquals($datum, $read_datum);
+        }
     }
-  }
 
-  public function test_differing_schemas_with_primitives()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-prim-%s.avr', $codec));
+    public function test_write_read_false_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-false-%s.avr', $codec));
+            $writers_schema = '"boolean"';
+            $datum = false;
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            $dw->append($datum);
+            $dw->close();
 
-      $writer_schema = <<<JSON
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $read_datum = array_shift($read_data);
+            $dr->close();
+            $this->assertEquals($datum, $read_datum);
+        }
+    }
+
+    public function test_write_read_int_array_round_trip()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-wr-int-ary-%s.avr', $codec));
+            $writers_schema = '"int"';
+            $data = array(10, 20, 30, 40, 50, 60, 70);
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            foreach ($data as $datum) {
+                $dw->append($datum);
+            }
+            $dw->close();
+
+            $dr = AvroDataIO::open_file($data_file);
+            $read_data = $dr->data();
+            $dr->close();
+            $this->assertEquals($data, $read_data,
+                sprintf("in: %s\nout: %s",
+                    json_encode($data), json_encode($read_data)));
+        }
+    }
+
+    public function test_differing_schemas_with_primitives()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-prim-%s.avr', $codec));
+
+            $writer_schema = <<<JSON
 { "type": "record",
   "name": "User",
   "fields" : [
@@ -203,35 +185,35 @@ class DataFileTest extends PHPUnit\Framework\TestCase
       {"name": "verified", "type": "boolean", "default": "false"}
       ]}
 JSON;
-      $data = array(array('username' => 'john', 'age' => 25, 'verified' => true),
-                    array('username' => 'ryan', 'age' => 23, 'verified' => false));
-      $dw = AvroDataIO::open_file($data_file, 'w', $writer_schema, $codec);
-      foreach ($data as $datum)
-      {
-        $dw->append($datum);
-      }
-      $dw->close();
-      $reader_schema = <<<JSON
+            $data = array(
+                array('username' => 'john', 'age' => 25, 'verified' => true),
+                array('username' => 'ryan', 'age' => 23, 'verified' => false)
+            );
+            $dw = AvroDataIO::open_file($data_file, 'w', $writer_schema, $codec);
+            foreach ($data as $datum) {
+                $dw->append($datum);
+            }
+            $dw->close();
+            $reader_schema = <<<JSON
         { "type": "record",
           "name": "User",
           "fields" : [
         {"name": "username", "type": "string"}
         ]}
 JSON;
-      $dr = AvroDataIO::open_file($data_file, 'r', $reader_schema);
-      foreach ($dr->data() as $index => $record)
-      {
-        $this->assertEquals($data[$index]['username'], $record['username']);
-      }
+            $dr = AvroDataIO::open_file($data_file, 'r', $reader_schema);
+            foreach ($dr->data() as $index => $record) {
+                $this->assertEquals($data[$index]['username'], $record['username']);
+            }
+        }
     }
-  }
 
-  public function test_differing_schemas_with_complex_objects()
-  {
-    foreach (AvroDataIO::valid_codecs() as $codec) {
-      $data_file = $this->add_data_file(sprintf('data-complex-%s.avr', $codec));
+    public function test_differing_schemas_with_complex_objects()
+    {
+        foreach (AvroDataIO::valid_codecs() as $codec) {
+            $data_file = $this->add_data_file(sprintf('data-complex-%s.avr', $codec));
 
-      $writers_schema = <<<JSON
+            $writers_schema = <<<JSON
 { "type": "record",
   "name": "something",
   "fields": [
@@ -251,40 +233,81 @@ JSON;
 ]}
 JSON;
 
-      $data = array(array("username" => "john",
-                          "something_fixed" => "foo",
-                          "something_enum" => "hello",
-                          "something_array" => array(1,2,3),
-                          "something_map" => array("a" => 1, "b" => 2),
-                          "something_record" => array("inner" => 2),
-                          "something_error" => array("code" => 403)),
-                    array("username" => "ryan",
-                          "something_fixed" => "bar",
-                          "something_enum" => "goodbye",
-                          "something_array" => array(1,2,3),
-                          "something_map" => array("a" => 2, "b" => 6),
-                          "something_record" => array("inner" => 1),
-                          "something_error" => array("code" => 401)));
-      $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
-      foreach ($data as $datum)
-        $dw->append($datum);
-      $dw->close();
+            $data = [
+                [
+                    "username" => "john",
+                    "something_fixed" => "foo",
+                    "something_enum" => "hello",
+                    "something_array" => [1, 2, 3],
+                    "something_map" => ["a" => 1, "b" => 2],
+                    "something_record" => ["inner" => 2],
+                    "something_error" => ["code" => 403]
+                ],
+                [
+                    "username" => "ryan",
+                    "something_fixed" => "bar",
+                    "something_enum" => "goodbye",
+                    "something_array" => [1, 2, 3],
+                    "something_map" => ["a" => 2, "b" => 6],
+                    "something_record" => ["inner" => 1],
+                    "something_error" => ["code" => 401]
+                ]
+            ];
+            $dw = AvroDataIO::open_file($data_file, 'w', $writers_schema, $codec);
+            foreach ($data as $datum) {
+                $dw->append($datum);
+            }
+            $dw->close();
 
-      foreach (array('fixed', 'enum', 'record', 'error',
-                     'array' , 'map', 'union') as $s)
-      {
-        $readers_schema = json_decode($writers_schema, true);
-        $dr = AvroDataIO::open_file($data_file, 'r', json_encode($readers_schema));
-        foreach ($dr->data() as $idx => $obj)
-        {
-          foreach ($readers_schema['fields'] as $field)
-          {
-            $field_name = $field['name'];
-            $this->assertEquals($data[$idx][$field_name], $obj[$field_name]);
-          }
+            foreach (array(
+                         'fixed',
+                         'enum',
+                         'record',
+                         'error',
+                         'array',
+                         'map',
+                         'union'
+                     ) as $s) {
+                $readers_schema = json_decode($writers_schema, true);
+                $dr = AvroDataIO::open_file($data_file, 'r', json_encode($readers_schema));
+                foreach ($dr->data() as $idx => $obj) {
+                    foreach ($readers_schema['fields'] as $field) {
+                        $field_name = $field['name'];
+                        $this->assertEquals($data[$idx][$field_name], $obj[$field_name]);
+                    }
+                }
+                $dr->close();
+            }
         }
-        $dr->close();
-      }
     }
-  }
+
+    protected function setUp(): void
+    {
+        if (!file_exists(TEST_TEMP_DIR)) {
+            mkdir(TEST_TEMP_DIR);
+        }
+        $this->remove_data_files();
+    }
+
+    protected function remove_data_files()
+    {
+        if (self::REMOVE_DATA_FILES
+            && !empty($this->data_files)) {
+            foreach ($this->data_files as $data_file) {
+                self::remove_data_file($data_file);
+            }
+        }
+    }
+
+    protected static function remove_data_file($data_file)
+    {
+        if (file_exists($data_file)) {
+            unlink($data_file);
+        }
+    }
+
+    protected function tearDown(): void
+    {
+        $this->remove_data_files();
+    }
 }
