+++ lang/php/test/InterOpTest.php
@@ -17,55 +17,61 @@
  * limitations under the License.
  */
 
-require_once('test_helper.php');
+namespace Apache\Avro\Tests;
 
-class InterOpTest extends PHPUnit\Framework\TestCase
+use Apache\Avro\DataFile\AvroDataIO;
+use Apache\Avro\IO\AvroFile;
+use Apache\Avro\Schema\AvroSchema;
+use PHPUnit\Framework\TestCase;
+
+class InterOpTest extends TestCase
 {
-  var $projection_json;
-  var $projection;
+    private $projection_json;
+    private $projection;
 
-  public function setUp(): void
-  {
-    $interop_schema_file_name = AVRO_INTEROP_SCHEMA;
-    $this->projection_json = file_get_contents($interop_schema_file_name);
-    $this->projection = AvroSchema::parse($this->projection_json);
-  }
+    public function setUp(): void
+    {
+        $interop_schema_file_name = AVRO_INTEROP_SCHEMA;
+        $this->projection_json = file_get_contents($interop_schema_file_name);
+        $this->projection = AvroSchema::parse($this->projection_json);
+    }
 
-  public function file_name_provider()
-  {
-    $data_dir = AVRO_BUILD_DATA_DIR;
-    $data_files = array();
-    if (!($dh = opendir($data_dir)))
-      die("Could not open data dir '$data_dir'\n");
+    public function file_name_provider()
+    {
+        $data_dir = AVRO_BUILD_DATA_DIR;
+        $data_files = array();
+        if (!($dh = opendir($data_dir))) {
+            die("Could not open data dir '$data_dir'\n");
+        }
 
-    while ($file = readdir($dh))
-      if (0 < preg_match('/^[a-z]+(_deflate|_snappy|_zstandard)?\.avro$/', $file))
-        $data_files []= implode(DIRECTORY_SEPARATOR, array($data_dir, $file));
-    closedir($dh);
+        while ($file = readdir($dh)) {
+            if (0 < preg_match('/^[a-z]+(_deflate|_snappy|_zstandard)?\.avro$/', $file)) {
+                $data_files [] = implode(DIRECTORY_SEPARATOR, array($data_dir, $file));
+            }
+        }
+        closedir($dh);
 
-    $ary = array();
-    foreach ($data_files as $df)
-      $ary []= array($df);
-    return $ary;
-  }
+        $ary = array();
+        foreach ($data_files as $df) {
+            $ary [] = array($df);
+        }
+        return $ary;
+    }
 
-  /**
-   *  @dataProvider file_name_provider
-   */
-  public function test_read($file_name)
-  {
+    /**
+     * @dataProvider file_name_provider
+     */
+    public function test_read($file_name)
+    {
+        $dr = AvroDataIO::open_file(
+            $file_name, AvroFile::READ_MODE, $this->projection_json);
 
-    $dr = AvroDataIO::open_file(
-      $file_name, AvroFile::READ_MODE, $this->projection_json);
+        $data = $dr->data();
 
-    $data = $dr->data();
+        $this->assertNotCount(0, $data, sprintf("no data read from %s", $file_name));
 
-    $this->assertNotEquals(0, count($data),
-                           sprintf("no data read from %s", $file_name));
-
-    foreach ($data as $idx => $datum)
-      $this->assertNotNull($datum, sprintf("null datum from %s", $file_name));
-
-  }
-
+        foreach ($data as $idx => $datum) {
+            $this->assertNotNull($datum, sprintf("null datum from %s", $file_name));
+        }
+    }
 }
