+++ lang/php/test/ProtocolFileTest.php
@@ -17,31 +17,18 @@
  * limitations under the License.
  */
 
-require_once('test_helper.php');
+namespace Apache\Avro\Tests;
 
 // near-verbatim port of test_protocol.py
-class ProtocolFileTest extends PHPUnit\Framework\TestCase
+use Apache\Avro\Protocol\AvroProtocol;
+use Apache\Avro\Schema\AvroSchemaParseException;
+use PHPUnit\Framework\TestCase;
+
+class ProtocolFileTest extends TestCase
 {
-	protected function setUp(): void {
-	}
-	
-	public function testParsing() {
-		$cnt=count($this->prot_parseable);
-		for ($i=0; $i<$cnt; $i++) {
-			try {
-				//print($i . " " . ($this->prot_parseable[$i]?"true":"false") . " \n");
-				$prot=AvroProtocol::parse($this->prot_data[$i]);
-			} catch (AvroSchemaParseException $x) {
-				// exception ok if we expected this protocol spec to be unparseable
-				$this->assertEquals(false, $this->prot_parseable[$i]);
-			}
-		}
-	}
-	
-	// test data
-	private $prot_parseable=array(true, true, true, true, true, true, false, true, true);
-	private $prot_data = array(
-<<<'DATUM'
+    private $prot_parseable = array(true, true, true, true, true, true, false, true, true);
+    private $prot_data = array(
+        <<<'DATUM'
 {
   "namespace": "com.acme",
   "protocol": "HelloWorld",
@@ -62,8 +49,8 @@ class ProtocolFileTest extends PHPUnit\Framework\TestCase
   }
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test",
  "protocol": "Simple",
 
@@ -118,8 +105,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test.namespace",
  "protocol": "TestNamespace",
 
@@ -149,8 +136,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test.namespace",
  "protocol": "TestImplicitNamespace",
 
@@ -184,8 +171,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test.namespace",
  "protocol": "TestNamespaceTwo",
 
@@ -222,8 +209,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test.namespace",
  "protocol": "TestValidRepeatedName",
 
@@ -256,8 +243,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test.namespace",
  "protocol": "TestInvalidRepeatedName",
 
@@ -289,8 +276,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {"namespace": "org.apache.avro.test",
  "protocol": "BulkData",
 
@@ -312,8 +299,8 @@ DATUM
 
 }
 DATUM
-,
-<<<'DATUM'
+    ,
+        <<<'DATUM'
 {
   "protocol" : "API",
   "namespace" : "xyz.api",
@@ -349,5 +336,21 @@ DATUM
   }
 }
 DATUM
-	);
+    );
+
+    // test data
+
+    public function testParsing()
+    {
+        $cnt = count($this->prot_parseable);
+        for ($i = 0; $i < $cnt; $i++) {
+            try {
+                //print($i . " " . ($this->prot_parseable[$i]?"true":"false") . " \n");
+                $prot = AvroProtocol::parse($this->prot_data[$i]);
+            } catch (AvroSchemaParseException $x) {
+                // exception ok if we expected this protocol spec to be unparseable
+                $this->assertEquals(false, $this->prot_parseable[$i]);
+            }
+        }
+    }
 }
