+++ lang/php/test/StringIOTest.php
@@ -17,69 +17,77 @@
  * limitations under the License.
  */
 
-require_once('test_helper.php');
+namespace Apache\Avro\Tests;
 
-class StringIOTest extends PHPUnit\Framework\TestCase
+use Apache\Avro\AvroDebug;
+use Apache\Avro\AvroIO;
+use Apache\Avro\DataFile\AvroDataIOReader;
+use Apache\Avro\DataFile\AvroDataIOWriter;
+use Apache\Avro\Datum\AvroIODatumReader;
+use Apache\Avro\Datum\AvroIODatumWriter;
+use Apache\Avro\IO\AvroStringIO;
+use Apache\Avro\Schema\AvroSchema;
+use PHPUnit\Framework\TestCase;
+
+class StringIOTest extends TestCase
 {
+    public function test_write()
+    {
+        $strio = new AvroStringIO();
+        $this->assertEquals(0, $strio->tell());
+        $str = 'foo';
+        $strlen = strlen($str);
+        $this->assertEquals($strlen, $strio->write($str));
+        $this->assertEquals($strlen, $strio->tell());
+    }
 
-  public function test_write()
-  {
-    $strio = new AvroStringIO();
-    $this->assertEquals(0, $strio->tell());
-    $str = 'foo';
-    $strlen = strlen($str);
-    $this->assertEquals($strlen, $strio->write($str));
-    $this->assertEquals($strlen, $strio->tell());
-  }
+    public function test_seek()
+    {
+        $strio = new AvroStringIO('abcdefghijklmnopqrstuvwxyz');
+        $strio->seek(4, AvroIO::SEEK_SET);
+        $this->assertEquals('efgh', $strio->read(4));
+        $strio->seek(4, AvroIO::SEEK_CUR);
+        $this->assertEquals('mnop', $strio->read(4));
+        $strio->seek(-4, AvroIO::SEEK_END);
+        $this->assertEquals('wxyz', $strio->read(4));
+    }
 
-  public function test_seek()
-  {
-    $strio = new AvroStringIO('abcdefghijklmnopqrstuvwxyz');
-    $strio->seek(4, AvroIO::SEEK_SET);
-    $this->assertEquals('efgh', $strio->read(4));
-    $strio->seek(4, AvroIO::SEEK_CUR);
-    $this->assertEquals('mnop', $strio->read(4));
-    $strio->seek(-4, AvroIO::SEEK_END);
-    $this->assertEquals('wxyz', $strio->read(4));
-  }
+    public function test_tell()
+    {
+        $strio = new AvroStringIO('foobar');
+        $this->assertEquals(0, $strio->tell());
+        $strlen = 3;
+        $strio->read($strlen);
+        $this->assertEquals($strlen, $strio->tell());
+    }
 
-  public function test_tell()
-  {
-    $strio = new AvroStringIO('foobar');
-    $this->assertEquals(0, $strio->tell());
-    $strlen = 3;
-    $strio->read($strlen);
-    $this->assertEquals($strlen, $strio->tell());
-  }
+    public function test_read()
+    {
+        $str = 'foobar';
+        $strio = new AvroStringIO($str);
+        $strlen = 3;
+        $this->assertEquals(substr($str, 0, $strlen), $strio->read($strlen));
+    }
 
-  public function test_read()
-  {
-    $str = 'foobar';
-    $strio = new AvroStringIO($str);
-    $strlen = 3;
-    $this->assertEquals(substr($str, 0, $strlen), $strio->read($strlen));
-  }
+    public function test_string_rep()
+    {
+        $writers_schema_json = '"null"';
+        $writers_schema = AvroSchema::parse($writers_schema_json);
+        $datum_writer = new AvroIODatumWriter($writers_schema);
+        $strio = new AvroStringIO();
+        $this->assertEquals('', $strio->string());
+        $dw = new AvroDataIOWriter($strio, $datum_writer, $writers_schema_json);
+        $dw->close();
 
-  public function test_string_rep()
-  {
-    $writers_schema_json = '"null"';
-    $writers_schema = AvroSchema::parse($writers_schema_json);
-    $datum_writer = new AvroIODatumWriter($writers_schema);
-    $strio = new AvroStringIO();
-    $this->assertEquals('', $strio->string());
-    $dw = new AvroDataIOWriter($strio, $datum_writer, $writers_schema_json);
-    $dw->close(); 
+        $this->assertEquals(57, strlen($strio->string()),
+            AvroDebug::ascii_string($strio->string()));
 
-    $this->assertEquals(57, strlen($strio->string()), 
-                        AvroDebug::ascii_string($strio->string()));
+        $read_strio = new AvroStringIO($strio->string());
 
-    $read_strio = new AvroStringIO($strio->string());
-
-    $datum_reader = new AvroIODatumReader();
-    $dr = new AvroDataIOReader($read_strio, $datum_reader);
-    $read_data = $dr->data();
-    $datum_count = count($read_data);
-    $this->assertEquals(0, $datum_count);
-  }
-
+        $datum_reader = new AvroIODatumReader();
+        $dr = new AvroDataIOReader($read_strio, $datum_reader);
+        $read_data = $dr->data();
+        $datum_count = count($read_data);
+        $this->assertEquals(0, $datum_count);
+    }
 }
